requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Gmail API dependencies
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0

# Utility dependencies
python-dateutil==2.8.2
python-dotenv==1.0.0

# Gemini AI dependencies
google-generativeai>=0.8.3

# OpenAI dependencies
openai>=1.0.0
tiktoken>=0.5.0

# Anthropic dependencies
anthropic>=0.21.0

# Development dependencies (optional)
pytest==7.4.4
black==23.12.1
flake8==7.0.0